refactor: move non-interactive project scoping into lib/detection
Detection logic lives with the other detectors, not in the programs
folder: posthog-integration/detect-agentic.ts becomes
lib/detection/project-scope.ts (content unchanged), next to agentic.ts
and framework.ts. The program keeps only the one-line ciPreRun call.

Generated-By: PostHog Code
Task-Id: 06aaaf67-5aba-497e-bd9e-0220e2417442
